Issue 30894 validating content type with two unique fields is not being handled (#31108)
The bug occurs because when an exception is thrown while executing a
statement in PostgreSQL, the transaction is automatically aborted.

So when you have more than one unique fields in the same Content Type
and the first field fail, when we try to use the same Transaction to
validate the second unique fields then you got the Abort Transaction
Error.
 
So we are going to stop the validation of the others fields when the
first unique field fail.

### Proposed Changes
* Throw Exception to stop validation when unique fields validation fail
by duplicated


https://github.com/dotCMS/core/pull/31108/files#diff-fa1ceaa19618a6b2bbc30e24c6f930b4971f417db50babb748c2e2837ba9eb82R7696

* We need to set hasError to true if any other error occurs


https://github.com/dotCMS/core/pull/31108/files#diff-fa1ceaa19618a6b2bbc30e24c6f930b4971f417db50babb748c2e2837ba9eb82R7700

* make sure that the Transaction is roolback if a duplication error was
thrown


https://github.com/dotCMS/core/pull/31108/files#diff-285b5a1d0f200a885dd7ad9f06a3c3188cfe2a3f580f676dffbfaab891aa1bf6R2382