feat(deployment): #31193 Added conditional logic to detect Kubernetes probes. (#31235)
## Proposed Changes

* This pull request includes an important update to the
`AppContextInitResource` class to handle Kubernetes probes more
efficiently. The change introduces a conditional check for the
User-Agent header to detect Kubernetes probes and respond with a
lightweight 200 OK response, avoiding the unnecessary payload of
configuration data.

**Key changes:**

* Added logic to detect Kubernetes probes by checking if the User-Agent
header starts with `kube-probe` and respond with a `200 OK` status
without the full configuration payload. This is a temporary solution to
handle probes efficiently until a dedicated endpoint is implemented.

### Fixes
#31193