fix(TimeMachine) Permissionable shortLivedKey was bad + Skip Velocity cache when time-machine is set (#31222)
### Proposed Changes
* I'm changing the shortLivedKey used to cache permissions for
Contentlets, the problem we want to fix here is that in the permissions
cache, we can have an identifier saying that a user did not have
permissions on a working and live version, while the user, in reality,
had permissions on the live version.
* Ultimately, I introduced a change to skip caching the velocity
templates when the future time machine runs. This is a cleaner fix for
the problem of unmatching content shown in both sides, the rendered
portion and the headless pieces.