chore(FTM): Resolve feedback found on IQA (#31266)
This pull request includes several updates to the `dot-uve-toolbar` and
`dot-uve-device-selector` components in the `edit-ema-editor` module.
The main changes focus on modifying date handling, updating tooltips,
and enhancing test coverage.

### Changes to Date Handling:
*
[`dot-uve-toolbar.component.html`](diffhunk://#diff-9937556e73b051b878ba22ad1ce971a70019a617d7979b3e0bcc814801ad350bL23-R24):
Changed the `minDate` attribute to use `MIN_DATE` and added
`dataType="date"` to the calendar component.
*
[`dot-uve-toolbar.component.ts`](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L131-R141):
Added a `MIN_DATE` getter that sets the time to 0 hours, 0 minutes, 0
seconds, and 0 milliseconds to ensure the minimum date does not collide
with the preview date.

### Tooltip and Button Updates:
*
[`dot-uve-device-selector.component.html`](diffhunk://#diff-13b3f8a3ba001e8aac1ac47383a943b96b6af0c99c43aa91154da1cf1be5d6a6L37-L38):
Removed the tooltip attribute and its position from the "more" button.

### Test Enhancements:
*
[`dot-uve-toolbar.component.spec.ts`](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1R624):
Added a mock date and a test to verify the `minDate` attribute of the
calendar component.
[[1]](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1R624)
[[2]](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1R671-R681)
*
[`dot-uve-device-selector.component.spec.ts`](diffhunk://#diff-652f846aa914e51df03471614a9ffc9b80757f275a3639fe1131def21c40b39dL62):
Removed the `setSocialMedia` mock function from the base UVE state.
*
[`dot-uve-toolbar.component.spec.ts`](diffhunk://#diff-3eaa147616a5d1ff374a5fa27b0f38f0159a9039ef7e8d672dec43631f48a9e1R58):
Added a mock date initialization in the `describe` block.

### Code Refactoring:
*
[`dot-uve-toolbar.component.ts`](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L119-R123):
Refactored the `$previewDate` computed property to use a more readable
function block.