Issue 31101 Copy url in experiments fails (#31275)
The url that was being saved in a variant was only the path to the page,
so a new url in the copy button was added. This url is processed and
made by a new function inside the component, that gets the current
location and create a whole url with the host also.

Now the url copied is for example
http://localhost:8082/blog/post/french-polynesia-everything-you-need-to-know?language_id=1&com.dotmarketing.persona.id=modes.persona.no.persona&editorMode=edit&disabledNavigateMode=true&mode=LIVE&variantName=DEFAULT

This url should open the page in the specified variant live version of
the page as client, just like clicking on _"Open Published Version"_.