fix(SDK): Update SDK Library and UVE search bar to Use "personaId" Key (#31307)
This pull request includes changes to the `DotPageApiService` and
related files to improve the handling of page modes based on the editor
mode. The main changes involve introducing a new enum `PAGE_MODE` and
updating the test cases and utility functions accordingly.

### Changes to `DotPageApiService`:

* Added import for `PAGE_MODE` in `dot-page-api.service.spec.ts` to use
the new enum for page modes.
* Added a new test case `editMode` to verify that the correct
`PAGE_MODE` is requested based on the `editorMode`.
* Updated existing test cases to use `PAGE_MODE.PREVIEW` and
`PAGE_MODE.LIVE` instead of hardcoded strings.
[[1]](diffhunk://#diff-1dad8c8d685da624eaecb75a9607377dfa7168120d182df479ffceaed1a87858L146-R167)
[[2]](diffhunk://#diff-1dad8c8d685da624eaecb75a9607377dfa7168120d182df479ffceaed1a87858L164-R185)

### Changes to utility functions:

* Added import for `UVE_MODE_TO_PAGE_MODE` and `PAGE_MODE` in `index.ts`
to map editor modes to page modes.
* Updated `getFullPageURL` function to use `UVE_MODE_TO_PAGE_MODE` for
setting the `mode` parameter based on `editorMode`.

### Video



https://github.com/user-attachments/assets/341589f0-1f26-4623-a516-faeba77c9b9c