fix(uve): Fixed double touch and color in FTM p-calendar (#31361)
https://github.com/user-attachments/assets/36df8c72-ec8b-4ef0-a7a3-b696e9bd96d6



This pull request includes changes to improve the styling and
functionality of the `dot-uve-toolbar` component and update the calendar
component's theme. The most important changes include updating the
background color for the selected calendar date, modifying the minimum
date handling in the `dot-uve-toolbar` component, and adding a new
method to get the minimum date.

Styling updates:

*
[`core-web/libs/dotcms-scss/angular/dotcms-theme/components/form/_calendar.scss`](diffhunk://#diff-275e3842fd0804fa57ecc4febce1fd83b8210b193643d925d1c187ac3cfa28bfL101-R101):
Updated the background color for the selected date in the calendar
component to use the primary color palette with 10% opacity.

Functionality improvements:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/dot-uve-toolbar.component.html`](diffhunk://#diff-9937556e73b051b878ba22ad1ce971a70019a617d7979b3e0bcc814801ad350bL24-R24):
Changed the `minDate` binding to use the `$MIN_DATE()` method instead of
the `MIN_DATE` property.
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/dot-uve-toolbar.component.ts`](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L13-R14):
Added the `signal` import and replaced the `MIN_DATE` getter with the
`$MIN_DATE` signal.
[[1]](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L13-R14)
[[2]](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30L133-R134)
*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/edit-ema-editor/components/dot-uve-toolbar/dot-uve-toolbar.component.ts`](diffhunk://#diff-217a9e619d6590c4f652e85353b9637ba5e464ddeb0424be35aef39bb8dceb30R302-R319):
Added a private `#getMinDate` method to encapsulate the logic for
calculating the minimum date.

Co-authored-by: Kevin Davila <56242609+kevindaviladev@users.noreply.github.com>