fix(CICD): Pinning tonistiigi/binfmt version to qemu-v7.0.0-28 (#31376) (#31377)
### Proposed Changes

* This pull request includes updates to the Docker setup in the CI/CD
pipeline to use a specific version of the QEMU binary format image. The
most important changes are:

Updates to Docker setup:

* Changed the QEMU image version from `tonistiigi/binfmt:latest` to
`tonistiigi/binfmt:qemu-v7.0.0-28` to ensure consistency and avoid
potential issues with the latest version.
* Updated the QEMU image version from `tonistiigi/binfmt:latest` to
`tonistiigi/binfmt:qemu-v7.0.0-28` to match the change in the deployment
action and maintain uniformity across the CI/CD pipeline.

### Fixes
#31376