fix(graphql) make graphql accept GETs (#31396)
This pull request introduces several changes to the
`DotGraphQLHttpServlet` class in order to improve the handling of
GraphQL requests. The key modifications include adding a new request
wrapper class and updating the `doGet` method to enforce the presence of
a query id `?qid=abc123` parameter that can be used to enforce
uniqueness in the responses.


ref: #31395