fix(pageAPI): Apply Regex and sanitization correctly (#31422)
### Problems found

1. Regarding the first mentioned issue, the vanityUrl was trying to
redirect to $1, this is because when handling regex in vanityUrl a
function was needed to be executed to process the url and get the final
one that the page should be redirected to, and in the cases of Temporary
and Permanent Redirect that function was not being executed
2. About the second issue, it was a case that was not being contemplated
when using vanityUrls. As the uri is `(.*)/index` and the forwardTo `$1`
this resulted in an empty string url that was generating a blank page
when accesing through browser.
3. Also the regex of the second issue was generating an infinite loop in
the UVE, this was because it was trying to forward to an empty string.
The frontend of the UVE was handling this case and fallbacking to
`index` in the url, then navigating to pageAPI to render `index` but
then the backend finds that there is a vanityUrl and forwards to empty
string, then the frontend redirects again to `index` and so on.

### Proposed Changes
* To solve the problem **1**: Apply a workaround to modify the forwardTo
value (executing the function that processes the url) and set it to the
CachedVanityUrl which is the entity used in the Temporary and Permanent
Redirect case. This way in those cases the forwardTo is going to be
correct, and return to the frontend as it should.
* To solve the problem **2**: Add an extra check in the `processForward`
function that verifies if the forward url is empty (this means that the
forward should redirect to the root) then set `/` as the final url. This
will make the url work as it should in the browser.
* To solve the problem **3**: Modify what the frontend send to the
backend in the cases where it receives from the frontend an empty url.
Instead of index, send to the backend `/`, this will result in the same
behaviour as index, but allowing the user to handle this last case. Now
when using a vanityUrl with `index` in the regex, and forwarding to `/`
it won't generate an infinite loop.
* Additional fix: when using a vanityUrl with regex, and forwarding to
an url that matches with that regex, it results in an infinite loop
error. So instead of using `DotPreconditions.checkArgument` to throw an
error, I made a check to see if the url is self referenced, and if it
true, then fallback to the main url.


**Note and question:** `final String urlIn =
!url.startsWith(StringPool.SLASH) ? StringPool.SLASH + url : url;` Was
added in the processForward function, this was because as navigating
through the UVE, the url that the backend gets doesn't have any starting
slashes, so in the case that the issue specifies, using `(.*)/index`
regex (and forwardTo $1) didn't match when navigating to index as the
url that the function receives would be `index` (without slash). So as
the `resolveVanityUrlIfPresent` does, this will try to check for matches
on the vanityUrl but with a `correctedUri` that will have a `/` at the
start. Before the addition of the slash and as the url didn't matched,
the forwardTo field remained on $1 and then we got the same error as the
problem 1. Adding the slash resulted in the problem 3, which was also
solved.
So the questions would be if this is ok. Should we handle it in a
different way? Could be some cases in where the addition of the slash
results in a wrong url or redirect?

### Checklist
- [x] Tests

---------

Co-authored-by: Jalinson Diaz <zjaaaldev@gmail.com>