feat(Edit Mode) #30799 : Creating a REST Endpoint that abstract the Lucene Query generation and return results (#31439)
### Proposed Changes
* Provides a new REST Endpoint that allow users/developers to retrieve
content via Lucene from Elasticsearch.
* Uses the recently created `LuceneQueryBuilder` service to generate the
appropriate query, and passes it down to the API to retrieve contents.
* The new Endpoint re-uses the part of the code used by the old
`/api/content/_search` that:
  * Retrieves the results from Elasticsearch.
* Generates the View object with such results and additional search
metadata.
* The re-usable code was moved to the `ContentHelper` class. All classes
in the core project that used such methods were updated, but are NOT
part of the goal of this PR.
* Some code issues were found during the creation of the Postman tests.
The respective code fixes are included as well.