fix(permissions): respect frontend permissions. ref. #31511 (#31515)
This pull request includes changes to the `DetailPageTransformerImpl`
class in the `dotCMS` project. The changes primarily focus on enhancing
the handling of user roles and permissions when transforming content
type detail pages.

Key changes:

* **Respecting User Roles:**
* Updated the `uriToId` method to respect anonymous permissions based on
the current request's page mode. This ensures that the method respects
the roles and permissions of the user making the request.
* Updated the `idToUri` method similarly to respect anonymous
permissions, ensuring consistent behavior across both transformations.