fix(FTM): urlContentletMap attributes does not match page rendered in future live preview: #31402 (#31490)
This pull request includes several changes to the
`HTMLPageAssetRenderedBuilder` class to enhance the functionality
related to contentlet retrieval and Time Machine support.

Key changes include:
* Modified the `findUrlContentlet` method to include additional logic
for retrieving contentlets based on Time Machine dates. This includes
checking for `WIKI_CONTENTLET_INODE` and `WIKI_CONTENTLET` attributes
and finding the appropriate contentlet versions.
* Introduced a new private method `getContentletForTimeMachine` to
handle the logic of finding contentlet versions based on a specified
Time Machine date. This method ensures that the correct version of a
contentlet is retrieved if a Time Machine date is configured.

### Video


https://github.com/user-attachments/assets/99352b43-b91b-48ca-8bf1-d9117f60bd02

---------

Co-authored-by: valentinogiardino <valentinogiardino17@gmail.com>
Co-authored-by: Valentino Giardino <77643678+valentinogiardino@users.noreply.github.com>