#31457: Adding a better way of invalidating OpenAI models (#31566)
Introducing usage of `CircuitBreakerUrl` instead of custom HTTP client
wrapper to provide a more robust way of consuming the OpenAI API.
Better way of identifying errors from OpenAI to handle and whether or
not to end up invalidating the model if it is justified enough.
Unit and ITs were updated.