fix(sdk): Error Throw outside UVE (#31585)
This pull request includes changes to the `createUVESubscription`
function and its related tests to improve error handling and logging.
The most important changes are as follows:

### Improvements to Error Handling and Logging:

*
[`core-web/libs/sdk/uve/src/lib/utils.ts`](diffhunk://#diff-9cb9e61b3ad482cd4dafebfdb3d19e42876b180919442cc0e19e0d4a5b30bab7L79-R103):
Modified `createUVESubscription` to log warnings and errors instead of
throwing exceptions when not running inside UVE or when an invalid event
is provided. The function now returns a subscription object with an
`unsubscribe` method that does nothing in these cases.

### Updates to Unit Tests:

*
[`core-web/libs/sdk/uve/src/lib/utils.spec.ts`](diffhunk://#diff-3171d3b9211959cd7faf576c2af2e689b837cbf14b375017e85fa03644d05fdaL270-R293):
Updated the test case for when not running inside UVE to check for a
logged warning instead of an error, and to verify the returned
subscription object.
*
[`core-web/libs/sdk/uve/src/lib/utils.spec.ts`](diffhunk://#diff-3171d3b9211959cd7faf576c2af2e689b837cbf14b375017e85fa03644d05fdaL297-R314):
Updated the test case for an invalid event to check for a logged error
instead of an exception, and to verify the returned subscription object.