fix(FTM): future contentlet image not updating correctly (#31590)
### Proposed Changes
* Enhanced `ShortyServlet.java` to execute Time Machine logic when
needed.
* Updated `PageResource.java` to adjust the moment session attributes
are removed from the session.
* Updated `withEditor.ts` to use `untracked` for safe access to store
parameters.

### Checklist
- [x] Tests
- [x] Translations
- [x] Security Implications Contemplated (add notes if applicable)

### Additional Info
`ShortyServlet` wasn't considering the retrieval of images by identifier
in the Time Machine context. Some requests were hitting `ShortyServlet`
after Time Machine cleanup was done in `PageResource`, resulting in
incorrect images being displayed. This pull request includes several
changes to improve the images rendered when using the Time Machine
feature.

### Time Machine Enhancements:
* Modified the `getPageRender` method to remove the reset of the Time
Machine in the `finally` block, and updated the
`setUpTimeMachineIfPresent` method to reset the Time Machine if not
needed.
* Added the `getTimeMachineContentlet` method to retrieve future
versions of contentlets if a Time Machine date is configured, and
updated the `getContentletByIdentifier` method to use this new method.

### URL Handling Improvements:
* During troubleshooting, we detected a double load of the iframe
causing double requests per image to `ShortyServlet` in the edit page.
Therefore, we updated the `withEditor` function to use the `untracked`
utility for accessing store parameters, ensuring that the parameters are
not tracked by the reactive system.

### Screenshots


https://github.com/user-attachments/assets/6b48185f-3192-4674-b7a9-2c30485b8055