fix(UVE): Respect App URL Configuration in the UVE (#31755)
This pull request includes changes to the
`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor`
module to enhance the handling of iframe URLs. The most important
changes are the addition of new test cases to verify the correct iframe
URL generation and the modification of the URL construction logic to
handle trailing slashes in the client host.

Enhancements to iframe URL handling:

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor/withEditor.spec.ts`](diffhunk://#diff-48915538475425e5656151bb2a73df7a99b4c80c2817c085f96352ef3252d5cbR260-R289):
Added new test cases to verify that the iframe URL is correctly set when
the `clientHost` is present, both with and without an additional path.

*
[`core-web/libs/portlets/edit-ema/portlet/src/lib/store/features/editor/withEditor.ts`](diffhunk://#diff-86e692578757ed7f4f6cba5d0aeb07641312f3b17885825d1a45987153ae87f0R58-R61):
Modified the `buildIframeURL` function to remove trailing slashes from
the `clientHost` before constructing the iframe URL. This ensures that
the URL is correctly formatted regardless of the presence of trailing
slashes.

### Video



https://github.com/user-attachments/assets/44993bab-2c44-425f-ab62-0218a047e9df