refactor(core): remove 'Parent Issue' field from defect, feature, and task templates (#31789)
### Proposed Changes

This pull request includes changes to the issue templates in the
`.github/ISSUE_TEMPLATE` directory. The main change is the removal of
the "Parent Issue" field from the defect, feature, and task templates.

Changes to issue templates:

*
[`.github/ISSUE_TEMPLATE/defect.yaml`](diffhunk://#diff-1fbf14bf5c9a66f985e0ddc2a78388f794294e49367c4ec1dc4872d566f5e4d3L12-L21):
Removed the "Parent Issue" field.
*
[`.github/ISSUE_TEMPLATE/feature.yaml`](diffhunk://#diff-6beb416586c4731c5c2194db41281446efceebf9f99cdbe2fa569d4a50e2d06fL13-L22):
Removed the "Parent Issue" field.
*
[`.github/ISSUE_TEMPLATE/task.yaml`](diffhunk://#diff-8a66f0993793dfea458176fce664e6d39e3be13d3c1aadf0e53225f4ca4f0a49L13-L22):
Removed the "Parent Issue" field.

### Checklist
- [ ] Tests
- [ ] Translations
- [ ] Security Implications Contemplated (add notes if applicable)