fix(Push Publish): Add query param to avoid wrong results on unique field search (#31795)
When push publishing new content with a unique field, the receiver was
verifying whether another contentlet with the same field value already
existed. If a match was found, the existing content was replaced.
However, the check did not consider whether the matching content
belonged to the same site.

As a result, if two contentlets had the same unique field value despite
being in different sites, one of them was incorrectly overwritten.

Fix
A new parameter (`host`) has been added to the Lucene query used for
this validation. This ensures that the check only considers content
within the same site before performing a replacement. Also this new
parameter is added only when the `uniquePerSite` variable is true,
because if it is false, then it doesn't matter which site is the content
from.

Now, push publishing will only overwrite content if an existing
contentlet with the same unique field is found within the same site,
preventing unintended overwrites across different sites.