fix: Checkbox filter not working as expected - fix conflict scenario (#31878)
As
[commented](https://github.com/dotCMS/core/issues/31649#issuecomment-2790530609)
in the issue, when the content with the relationship field had more than
one relationship to the same content type, an already existing id error
was occurring.

The fields were previously identificated with the `dialogCounter`
variable, so now we keep that on the id field, but we remove it from the
`name` one, this way replicating the Content Search portlet data
formatting and also keeping the consistency and difference between
fields.



https://github.com/user-attachments/assets/667fe9da-51bf-41e4-9117-d722db50783c