Feat (Core) Add experiment in the page field in GraphQL (#31862)
This pull request introduces a new feature to the GraphQL API by adding
support for fetching the identifier of a running experiment associated
with a page. The changes include updates to the
`PageAPIGraphQLTypesProvider` class, the addition of a new data fetcher
class, and corresponding tests.

### New Feature: Running Experiment Fetcher

* **Integration of `RunningExperimentFetcher` in GraphQL Types
Provider:**
- Added import for `RunningExperimentFetcher` in
`PageAPIGraphQLTypesProvider.java`.
- Added `runningExperimentId` field to the `pageFields` map in the
`getTypes` method of `PageAPIGraphQLTypesProvider.java`.

* **Implementation of `RunningExperimentFetcher` Class:**
- Created `RunningExperimentFetcher` class to fetch the identifier of a
running experiment for a page. This class interacts with the
`ExperimentsAPI` to retrieve the necessary data.

* **Unit Tests for `RunningExperimentFetcher`:**
- Added `RunningExperimentFetcherTest` class to test the functionality
of `RunningExperimentFetcher`. The tests cover scenarios where a page
has no running experiment and where a page has an active running
experiment.