test: fixing npe on host.getTagStroage (#31825)
This pull request includes changes to the `Host` class in the `dotCMS`
project, focusing on improving the handling of the `tagStorage`
attribute and adding relevant tests. The most important changes include
updating the `getTagStorage` method to use the `Try` class for error
handling and adding a new test method to ensure the updated
functionality works correctly.

### Improvements to `Host` class:

*
[`dotCMS/src/main/java/com/dotmarketing/beans/Host.java`](diffhunk://#diff-9b11e0803028bbd952585ed75c35becfef165bb32222dca4ed0b0ab06e201ca3L188-R196):
Updated the `getTagStorage` method to use `Try` for error handling and
return the host identifier or `Host.SYSTEM_HOST` if the host is not
found.
*
[`dotCMS/src/main/java/com/dotmarketing/beans/Host.java`](diffhunk://#diff-9b11e0803028bbd952585ed75c35becfef165bb32222dca4ed0b0ab06e201ca3R13):
Added import for `Try` from `io.vavr.control` to support the new error
handling mechanism.

### Additions to tests:

*
[`dotcms-integration/src/test/java/com/dotmarketing/beans/HostTest.java`](diffhunk://#diff-a27f8d6452af05e16631d9760ad975bc15d4144fdafd910bb7a024c90c4cbcdbR46-R73):
Added a new test method `insureTagStorageWorks` to verify the updated
`getTagStorage` functionality.
*
[`dotcms-integration/src/test/java/com/dotmarketing/beans/HostTest.java`](diffhunk://#diff-a27f8d6452af05e16631d9760ad975bc15d4144fdafd910bb7a024c90c4cbcdbR3-R9):
Added necessary imports for `Try` and `APILocator` to support the new
test method.ref: #31824