task(dev-image): build latest dev image nightly (#31820)
This pull request includes updates to the CI/CD workflow and adds a new
step to build and push a nightly `latest` Docker image for the
development environment.

Enhancements to CI/CD workflow:

*
[`.github/workflows/cicd_comp_deployment-phase.yml`](diffhunk://#diff-c10f22dc11cbbd6ec4cc28e94d9e83e204d995829f7a5c9de25bb5303496b03fR132-R152):
Added a new job to build and push a nightly Docker image for the
development environment, including necessary parameters and secrets.
*
[`.github/workflows/cicd_comp_deployment-phase.yml`](diffhunk://#diff-c10f22dc11cbbd6ec4cc28e94d9e83e204d995829f7a5c9de25bb5303496b03fL196):
Removed unnecessary whitespace for better readability.ref: #31819